Added a function that returns a GroupView based on the group id to the GroupSystem, and did some corresponding restructuring in GroupPage and ChatWindowPage.
MessangerGlueRemote.java
Renamed getRecentPostsXml to getPostsXml and changed its arguments to
be guid objects of user and post, as well as an optional outer element name for 
the stream of post-related objects.

Room.java
Have object xml for either post or group be included in the room info, when room details 
are requested.