Add optional flag to Cache.getSync/Cache.getAsync to force a new web service request. Pass this through to saveInCache, which can choose to for whatever reason not actually write new stuff to the cache. For the Flickr photo and photoset caches, don't save new results to the cache if the photo count is unchanged.
Rip the explicit expireCache() games out of Flickr and YouTube updaters, in favor of getSync(username, true) to force ignoring the cache.

expireCache is now "on notice" as something that should not really exist ... all cache management should be in the cache bean. The remaining couple of uses are somewhat involved to fix though.