Rename debugFeatures to serverFeatures to repurpose as a general-purpose feature toggle.
When pollingTask is enabled, FeedUpdaterPeriodicJob and LastFmUpdaterPeriodicJob
are disabled.

Annotate XmlMethodErrorCode with whether or not they are expected in
normal conditions.

Extend DynamicPollingSystem to support differentiation between 
normal exceptions and code errors.  The former are not logged
with stack traces.  

Remove task sets slower than 8 hours.

Add LastFm support for DynamicPollingSystem.