Create polling tasks for YouTube and Flickr, moving some infrastructure from LastFm subclass into CachedExternalUpdaterBean superclass. Note we still support the non-pollingTask feature mode but hopefully this will go away shortly.
PollingTaskPersistence now supports createTaskIdempotent which
we use in several places as required.  Specifically in all
task migration (we want migration methods to be idempotent).

ExternalAccountSystem now *always* calls the notification method
onExternalAccountLovedAndEnabledMaybeChanged when an account
is modified.  This is needed for the case where a user changes
their account handle, so we can create a polling task for it.
Callbacks for this method handle the "maybe" part of sentiment
regardless.

Add missing check for task change before bumping it to a faster
set.  This was improperly causing tasks just bumped to a slower
set to go back down to their old set because their periodicity
average was still low.

Log number of tasks which actually changed.

Now load new tasks every 5 seconds, but continue saving them every
30.  This lets us pick up on new external account tasks relatively
quickly.  Not quite a substitute for proper JMS notification but
good enough for now.

Cut down on DynamicPollingSystem log messages (no messages from
empty task sets or when no new tasks were loaded).