Use FacebookPhotoDataView to contain details about Facebook photos outside of the cache bean.
Don't keep a reference to CachedFacebookPhotoData in FacebookPhotoDataStatus.

Keep the Facebook photo id in FacebookPhotoDataStatus, don't use the encoded id and salt any 
longer, null them as soon as we match up the photo data we get from Facebook and are able 
to store the photo id.

Delete the cached photos explicitly only if someone's Facebook session expires.

Expire the cache if the number of photos the web service returns doesn't match the number of 
photos we have cached, not the number of photo data statuses, which might always be greater 
due to some removed photos.

Display the count associated with a past tagged photos event even if we don't have the actual 
photos available to us because we had to drop the cache or they were removed.