Removed the search info bar from the top area.  Still looking for a place to put the [clear] link if it needs to go back, but for now things look cleaner.
Removed the unused [_]getAction() code while I was in the area

Stripped out the existing searchSecondary system in favor of a sidebar element for the secondary searches.  Added the loading of the sidebar to the .html page.

Updated the (Ctrl-Enter) hint to be [Ctrl-Enter] in order to keep things uniform.

Moved the (search.length > 1) check to the handleSearchChanged function

Lots of CSS updates to clean up how the changes look