http://bugzilla.mugshot.org/show_bug.cgi?id=1309
This one I believe is fallout of some change in the DBus
GetBaseProperties call; we retrieved the "base url" and used
it to emit the "initialized" signal on the global mugshot.  But
that's not operating correctly when not logged in.

Simplest fix is to use a hardcoded a base url (online.gnome.org) if
none is specified.