Cleanups of google.py and the stocks using it... make it use accounts.py to know what Google accounts exist, and try to rationalize how authentication works.
The previous code assumed there was a separate login step, distinct from making the API calls.
So if the mail checker hadn't existed, we never would have logged in at all. The correct approach 
is to simply try making the API calls and fail appropriately if we don't authenticate when we do.

There is a lot left to do to get all this working reliably.