Do em.flush() after removing AcccountClaim.
Check if the Facebook external account is loved and enabled before trying to get 
a profile link for it.

Don't check termsOfUserAccepted to determine if the account is active, since if 
we have some information for the account it means it came from Facebook and therefore
should be returned or used to make requests.

Adjust the application page form some more.