Don't try to include invitations for resources who already have accounts claiming them in the list of invitations created by the inviter (this used to cause an exception). Such accounts are already included in the Network of the inviter.
Don't display contacts whose accounts are not public in the Network of the inviter.
Instead, display them in the "People I've Invited to Join Mugshot" section if their
is an explicit contact claim between the inviter and the contact's e-mail address.
Most likely the person has invited them at some point if there is such a contact claim.
Omit users with disabled accounts if such contact claim does not exist.