Save Gmail accounts as supported by Google by default.
Do not allow the user to uncheck them as such.

Signal a User data model change if an account claim 
is created with the user as an owner of an e-mail supported 
by Google. (This will apply if someone verified a Gmail
email, but also if someone claims an e-mail that was 
previously marked as supported by Google.)