Allow multiple accounts of the same type per user.
Create mugshotEnabled flag for ExternalAccount to indicate which accounts are going to be used by Mugshot (for now only one per account type).

Pass around external account id to know which account is being edited.

Change the way we display accounts on the web to be more generic in preparation to supporting entering multiple accounts.