AccountsIQHandler.java: Create an AccountsIQHandler with an addOnlineAccount method for adding online accounts. Add special handling for Google accounts that need to be verified.
ClaimVerifierBean.java: add an extra check to see if an e-mail is claimed by someone else and return an error to the user immediately instead of sending a verification e-mail.

IdentitySpiderBean.java: signal onGoogleServicedEmailChange if a Google e-mail was removed.

FeedSystemBean.java: fall back to using feed entry link as its guid if syndEntry.getUri() returns null, which can be the case if a feed entry is missing an optional guid element.

FacebookServlet.java: update the signature used when calling online accounts set methods.